Generics.SOP.Instances

Description

Instances for Generic and HasMetadata.

We define instances for datatypes from generics-sop and base that are supported.

(There are only instances defined in this module, so the documentation is empty.)

Orphan instances

HasDatatypeInfo Bool

Associated Types

type DatatypeInfoOf Bool :: DatatypeInfo

Methods

datatypeInfo :: proxy Bool -> DatatypeInfo (Code Bool)

HasDatatypeInfo Ordering

Associated Types

type DatatypeInfoOf Ordering :: DatatypeInfo

Methods

datatypeInfo :: proxy Ordering -> DatatypeInfo (Code Ordering)

HasDatatypeInfo ()

Associated Types

type DatatypeInfoOf () :: DatatypeInfo

Methods

datatypeInfo :: proxy () -> DatatypeInfo (Code ())

HasDatatypeInfo DataRep

Associated Types

type DatatypeInfoOf DataRep :: DatatypeInfo

Methods

datatypeInfo :: proxy DataRep -> DatatypeInfo (Code DataRep)

HasDatatypeInfo ConstrRep

Associated Types

type DatatypeInfoOf ConstrRep :: DatatypeInfo

Methods

datatypeInfo :: proxy ConstrRep -> DatatypeInfo (Code ConstrRep)

HasDatatypeInfo Fixity

Associated Types

type DatatypeInfoOf Fixity :: DatatypeInfo

Methods

datatypeInfo :: proxy Fixity -> DatatypeInfo (Code Fixity)

HasDatatypeInfo FormatAdjustment

Associated Types

type DatatypeInfoOf FormatAdjustment :: DatatypeInfo

Methods

datatypeInfo :: proxy FormatAdjustment -> DatatypeInfo (Code FormatAdjustment)

HasDatatypeInfo FormatSign

Associated Types

type DatatypeInfoOf FormatSign :: DatatypeInfo

Methods

datatypeInfo :: proxy FormatSign -> DatatypeInfo (Code FormatSign)

HasDatatypeInfo FieldFormat

Associated Types

type DatatypeInfoOf FieldFormat :: DatatypeInfo

Methods

datatypeInfo :: proxy FieldFormat -> DatatypeInfo (Code FieldFormat)

HasDatatypeInfo FormatParse

Associated Types

type DatatypeInfoOf FormatParse :: DatatypeInfo

Methods

datatypeInfo :: proxy FormatParse -> DatatypeInfo (Code FormatParse)

HasDatatypeInfo Version

Associated Types

type DatatypeInfoOf Version :: DatatypeInfo

Methods

datatypeInfo :: proxy Version -> DatatypeInfo (Code Version)

HasDatatypeInfo PatternMatchFail

Associated Types

type DatatypeInfoOf PatternMatchFail :: DatatypeInfo

Methods

datatypeInfo :: proxy PatternMatchFail -> DatatypeInfo (Code PatternMatchFail)

HasDatatypeInfo RecSelError

Associated Types

type DatatypeInfoOf RecSelError :: DatatypeInfo

Methods

datatypeInfo :: proxy RecSelError -> DatatypeInfo (Code RecSelError)

HasDatatypeInfo RecConError

Associated Types

type DatatypeInfoOf RecConError :: DatatypeInfo

Methods

datatypeInfo :: proxy RecConError -> DatatypeInfo (Code RecConError)

HasDatatypeInfo RecUpdError

Associated Types

type DatatypeInfoOf RecUpdError :: DatatypeInfo

Methods

datatypeInfo :: proxy RecUpdError -> DatatypeInfo (Code RecUpdError)

HasDatatypeInfo NoMethodError

Associated Types

type DatatypeInfoOf NoMethodError :: DatatypeInfo

Methods

datatypeInfo :: proxy NoMethodError -> DatatypeInfo (Code NoMethodError)

HasDatatypeInfo NonTermination

Associated Types

type DatatypeInfoOf NonTermination :: DatatypeInfo

Methods

datatypeInfo :: proxy NonTermination -> DatatypeInfo (Code NonTermination)

HasDatatypeInfo NestedAtomically

Associated Types

type DatatypeInfoOf NestedAtomically :: DatatypeInfo

Methods

datatypeInfo :: proxy NestedAtomically -> DatatypeInfo (Code NestedAtomically)

HasDatatypeInfo Errno

Associated Types

type DatatypeInfoOf Errno :: DatatypeInfo

Methods

datatypeInfo :: proxy Errno -> DatatypeInfo (Code Errno)

HasDatatypeInfo BlockedIndefinitelyOnMVar

Associated Types

type DatatypeInfoOf BlockedIndefinitelyOnMVar :: DatatypeInfo

Methods

datatypeInfo :: proxy BlockedIndefinitelyOnMVar -> DatatypeInfo (Code BlockedIndefinitelyOnMVar)

HasDatatypeInfo BlockedIndefinitelyOnSTM

Associated Types

type DatatypeInfoOf BlockedIndefinitelyOnSTM :: DatatypeInfo

Methods

datatypeInfo :: proxy BlockedIndefinitelyOnSTM -> DatatypeInfo (Code BlockedIndefinitelyOnSTM)

HasDatatypeInfo Deadlock

Associated Types

type DatatypeInfoOf Deadlock :: DatatypeInfo

Methods

datatypeInfo :: proxy Deadlock -> DatatypeInfo (Code Deadlock)

HasDatatypeInfo AssertionFailed

Associated Types

type DatatypeInfoOf AssertionFailed :: DatatypeInfo

Methods

datatypeInfo :: proxy AssertionFailed -> DatatypeInfo (Code AssertionFailed)

HasDatatypeInfo AsyncException

Associated Types

type DatatypeInfoOf AsyncException :: DatatypeInfo

Methods

datatypeInfo :: proxy AsyncException -> DatatypeInfo (Code AsyncException)

HasDatatypeInfo ArrayException

Associated Types

type DatatypeInfoOf ArrayException :: DatatypeInfo

Methods

datatypeInfo :: proxy ArrayException -> DatatypeInfo (Code ArrayException)

HasDatatypeInfo ExitCode

Associated Types

type DatatypeInfoOf ExitCode :: DatatypeInfo

Methods

datatypeInfo :: proxy ExitCode -> DatatypeInfo (Code ExitCode)

HasDatatypeInfo BufferMode

Associated Types

type DatatypeInfoOf BufferMode :: DatatypeInfo

Methods

datatypeInfo :: proxy BufferMode -> DatatypeInfo (Code BufferMode)

HasDatatypeInfo Newline

Associated Types

type DatatypeInfoOf Newline :: DatatypeInfo

Methods

datatypeInfo :: proxy Newline -> DatatypeInfo (Code Newline)

HasDatatypeInfo NewlineMode

Associated Types

type DatatypeInfoOf NewlineMode :: DatatypeInfo

Methods

datatypeInfo :: proxy NewlineMode -> DatatypeInfo (Code NewlineMode)

HasDatatypeInfo SeekMode

Associated Types

type DatatypeInfoOf SeekMode :: DatatypeInfo

Methods

datatypeInfo :: proxy SeekMode -> DatatypeInfo (Code SeekMode)

HasDatatypeInfo MaskingState

Associated Types

type DatatypeInfoOf MaskingState :: DatatypeInfo

Methods

datatypeInfo :: proxy MaskingState -> DatatypeInfo (Code MaskingState)

HasDatatypeInfo IOException

Associated Types

type DatatypeInfoOf IOException :: DatatypeInfo

Methods

datatypeInfo :: proxy IOException -> DatatypeInfo (Code IOException)

HasDatatypeInfo ErrorCall

Associated Types

type DatatypeInfoOf ErrorCall :: DatatypeInfo

Methods

datatypeInfo :: proxy ErrorCall -> DatatypeInfo (Code ErrorCall)

HasDatatypeInfo ArithException

Associated Types

type DatatypeInfoOf ArithException :: DatatypeInfo

Methods

datatypeInfo :: proxy ArithException -> DatatypeInfo (Code ArithException)

HasDatatypeInfo All

Associated Types

type DatatypeInfoOf All :: DatatypeInfo

Methods

datatypeInfo :: proxy All -> DatatypeInfo (Code All)

HasDatatypeInfo Any

Associated Types

type DatatypeInfoOf Any :: DatatypeInfo

Methods

datatypeInfo :: proxy Any -> DatatypeInfo (Code Any)

HasDatatypeInfo CChar

Associated Types

type DatatypeInfoOf CChar :: DatatypeInfo

Methods

datatypeInfo :: proxy CChar -> DatatypeInfo (Code CChar)

HasDatatypeInfo CSChar

Associated Types

type DatatypeInfoOf CSChar :: DatatypeInfo

Methods

datatypeInfo :: proxy CSChar -> DatatypeInfo (Code CSChar)

HasDatatypeInfo CUChar

Associated Types

type DatatypeInfoOf CUChar :: DatatypeInfo

Methods

datatypeInfo :: proxy CUChar -> DatatypeInfo (Code CUChar)

HasDatatypeInfo CShort

Associated Types

type DatatypeInfoOf CShort :: DatatypeInfo

Methods

datatypeInfo :: proxy CShort -> DatatypeInfo (Code CShort)

HasDatatypeInfo CUShort

Associated Types

type DatatypeInfoOf CUShort :: DatatypeInfo

Methods

datatypeInfo :: proxy CUShort -> DatatypeInfo (Code CUShort)

HasDatatypeInfo CInt

Associated Types

type DatatypeInfoOf CInt :: DatatypeInfo

Methods

datatypeInfo :: proxy CInt -> DatatypeInfo (Code CInt)

HasDatatypeInfo CUInt

Associated Types

type DatatypeInfoOf CUInt :: DatatypeInfo

Methods

datatypeInfo :: proxy CUInt -> DatatypeInfo (Code CUInt)

HasDatatypeInfo CLong

Associated Types

type DatatypeInfoOf CLong :: DatatypeInfo

Methods

datatypeInfo :: proxy CLong -> DatatypeInfo (Code CLong)

HasDatatypeInfo CULong

Associated Types

type DatatypeInfoOf CULong :: DatatypeInfo

Methods

datatypeInfo :: proxy CULong -> DatatypeInfo (Code CULong)

HasDatatypeInfo CLLong

Associated Types

type DatatypeInfoOf CLLong :: DatatypeInfo

Methods

datatypeInfo :: proxy CLLong -> DatatypeInfo (Code CLLong)

HasDatatypeInfo CULLong

Associated Types

type DatatypeInfoOf CULLong :: DatatypeInfo

Methods

datatypeInfo :: proxy CULLong -> DatatypeInfo (Code CULLong)

HasDatatypeInfo CFloat

Associated Types

type DatatypeInfoOf CFloat :: DatatypeInfo

Methods

datatypeInfo :: proxy CFloat -> DatatypeInfo (Code CFloat)

HasDatatypeInfo CDouble

Associated Types

type DatatypeInfoOf CDouble :: DatatypeInfo

Methods

datatypeInfo :: proxy CDouble -> DatatypeInfo (Code CDouble)

HasDatatypeInfo CPtrdiff

Associated Types

type DatatypeInfoOf CPtrdiff :: DatatypeInfo

Methods

datatypeInfo :: proxy CPtrdiff -> DatatypeInfo (Code CPtrdiff)

HasDatatypeInfo CSize

Associated Types

type DatatypeInfoOf CSize :: DatatypeInfo

Methods

datatypeInfo :: proxy CSize -> DatatypeInfo (Code CSize)

HasDatatypeInfo CWchar

Associated Types

type DatatypeInfoOf CWchar :: DatatypeInfo

Methods

datatypeInfo :: proxy CWchar -> DatatypeInfo (Code CWchar)

HasDatatypeInfo CSigAtomic

Associated Types

type DatatypeInfoOf CSigAtomic :: DatatypeInfo

Methods

datatypeInfo :: proxy CSigAtomic -> DatatypeInfo (Code CSigAtomic)

HasDatatypeInfo CClock

Associated Types

type DatatypeInfoOf CClock :: DatatypeInfo

Methods

datatypeInfo :: proxy CClock -> DatatypeInfo (Code CClock)

HasDatatypeInfo CTime

Associated Types

type DatatypeInfoOf CTime :: DatatypeInfo

Methods

datatypeInfo :: proxy CTime -> DatatypeInfo (Code CTime)

HasDatatypeInfo CUSeconds

Associated Types

type DatatypeInfoOf CUSeconds :: DatatypeInfo

Methods

datatypeInfo :: proxy CUSeconds -> DatatypeInfo (Code CUSeconds)

HasDatatypeInfo CSUSeconds

Associated Types

type DatatypeInfoOf CSUSeconds :: DatatypeInfo

Methods

datatypeInfo :: proxy CSUSeconds -> DatatypeInfo (Code CSUSeconds)

HasDatatypeInfo CIntPtr

Associated Types

type DatatypeInfoOf CIntPtr :: DatatypeInfo

Methods

datatypeInfo :: proxy CIntPtr -> DatatypeInfo (Code CIntPtr)

HasDatatypeInfo CUIntPtr

Associated Types

type DatatypeInfoOf CUIntPtr :: DatatypeInfo

Methods

datatypeInfo :: proxy CUIntPtr -> DatatypeInfo (Code CUIntPtr)

HasDatatypeInfo CIntMax

Associated Types

type DatatypeInfoOf CIntMax :: DatatypeInfo

Methods

datatypeInfo :: proxy CIntMax -> DatatypeInfo (Code CIntMax)

HasDatatypeInfo CUIntMax

Associated Types

type DatatypeInfoOf CUIntMax :: DatatypeInfo

Methods

datatypeInfo :: proxy CUIntMax -> DatatypeInfo (Code CUIntMax)

HasDatatypeInfo IOMode

Associated Types

type DatatypeInfoOf IOMode :: DatatypeInfo

Methods

datatypeInfo :: proxy IOMode -> DatatypeInfo (Code IOMode)

HasDatatypeInfo Lexeme

Associated Types

type DatatypeInfoOf Lexeme :: DatatypeInfo

Methods

datatypeInfo :: proxy Lexeme -> DatatypeInfo (Code Lexeme)

HasDatatypeInfo Number

Associated Types

type DatatypeInfoOf Number :: DatatypeInfo

Methods

datatypeInfo :: proxy Number -> DatatypeInfo (Code Number)

HasDatatypeInfo GeneralCategory

Associated Types

type DatatypeInfoOf GeneralCategory :: DatatypeInfo

Methods

datatypeInfo :: proxy GeneralCategory -> DatatypeInfo (Code GeneralCategory)

Generic Bool

Associated Types

type Code Bool :: [[*]]

Methods

from :: Bool -> Rep Bool

to :: Rep Bool -> Bool

Generic Ordering

Associated Types

type Code Ordering :: [[*]]

Methods

from :: Ordering -> Rep Ordering

to :: Rep Ordering -> Ordering

Generic ()

Associated Types

type Code () :: [[*]]

Methods

from :: () -> Rep ()

to :: Rep () -> ()

Generic DataRep

Associated Types

type Code DataRep :: [[*]]

Methods

from :: DataRep -> Rep DataRep

to :: Rep DataRep -> DataRep

Generic ConstrRep

Associated Types

type Code ConstrRep :: [[*]]

Methods

from :: ConstrRep -> Rep ConstrRep

to :: Rep ConstrRep -> ConstrRep

Generic Fixity

Associated Types

type Code Fixity :: [[*]]

Methods

from :: Fixity -> Rep Fixity

to :: Rep Fixity -> Fixity

Generic FormatAdjustment

Associated Types

type Code FormatAdjustment :: [[*]]

Methods

from :: FormatAdjustment -> Rep FormatAdjustment

to :: Rep FormatAdjustment -> FormatAdjustment

Generic FormatSign

Associated Types

type Code FormatSign :: [[*]]

Methods

from :: FormatSign -> Rep FormatSign

to :: Rep FormatSign -> FormatSign

Generic FieldFormat

Associated Types

type Code FieldFormat :: [[*]]

Methods

from :: FieldFormat -> Rep FieldFormat

to :: Rep FieldFormat -> FieldFormat

Generic FormatParse

Associated Types

type Code FormatParse :: [[*]]

Methods

from :: FormatParse -> Rep FormatParse

to :: Rep FormatParse -> FormatParse

Generic Version

Associated Types

type Code Version :: [[*]]

Methods

from :: Version -> Rep Version

to :: Rep Version -> Version

Generic PatternMatchFail

Associated Types

type Code PatternMatchFail :: [[*]]

Methods

from :: PatternMatchFail -> Rep PatternMatchFail

to :: Rep PatternMatchFail -> PatternMatchFail

Generic RecSelError

Associated Types

type Code RecSelError :: [[*]]

Methods

from :: RecSelError -> Rep RecSelError

to :: Rep RecSelError -> RecSelError

Generic RecConError

Associated Types

type Code RecConError :: [[*]]

Methods

from :: RecConError -> Rep RecConError

to :: Rep RecConError -> RecConError

Generic RecUpdError

Associated Types

type Code RecUpdError :: [[*]]

Methods

from :: RecUpdError -> Rep RecUpdError

to :: Rep RecUpdError -> RecUpdError

Generic NoMethodError

Associated Types

type Code NoMethodError :: [[*]]

Methods

from :: NoMethodError -> Rep NoMethodError

to :: Rep NoMethodError -> NoMethodError

Generic NonTermination

Associated Types

type Code NonTermination :: [[*]]

Methods

from :: NonTermination -> Rep NonTermination

to :: Rep NonTermination -> NonTermination

Generic NestedAtomically

Associated Types

type Code NestedAtomically :: [[*]]

Methods

from :: NestedAtomically -> Rep NestedAtomically

to :: Rep NestedAtomically -> NestedAtomically

Generic Errno

Associated Types

type Code Errno :: [[*]]

Methods

from :: Errno -> Rep Errno

to :: Rep Errno -> Errno

Generic BlockedIndefinitelyOnMVar

Associated Types

type Code BlockedIndefinitelyOnMVar :: [[*]]

Methods

from :: BlockedIndefinitelyOnMVar -> Rep BlockedIndefinitelyOnMVar

to :: Rep BlockedIndefinitelyOnMVar -> BlockedIndefinitelyOnMVar

Generic BlockedIndefinitelyOnSTM

Associated Types

type Code BlockedIndefinitelyOnSTM :: [[*]]

Methods

from :: BlockedIndefinitelyOnSTM -> Rep BlockedIndefinitelyOnSTM

to :: Rep BlockedIndefinitelyOnSTM -> BlockedIndefinitelyOnSTM

Generic Deadlock

Associated Types

type Code Deadlock :: [[*]]

Methods

from :: Deadlock -> Rep Deadlock

to :: Rep Deadlock -> Deadlock

Generic AssertionFailed

Associated Types

type Code AssertionFailed :: [[*]]

Methods

from :: AssertionFailed -> Rep AssertionFailed

to :: Rep AssertionFailed -> AssertionFailed

Generic AsyncException

Associated Types

type Code AsyncException :: [[*]]

Methods

from :: AsyncException -> Rep AsyncException

to :: Rep AsyncException -> AsyncException

Generic ArrayException

Associated Types

type Code ArrayException :: [[*]]

Methods

from :: ArrayException -> Rep ArrayException

to :: Rep ArrayException -> ArrayException

Generic ExitCode

Associated Types

type Code ExitCode :: [[*]]

Methods

from :: ExitCode -> Rep ExitCode

to :: Rep ExitCode -> ExitCode

Generic BufferMode

Associated Types

type Code BufferMode :: [[*]]

Methods

from :: BufferMode -> Rep BufferMode

to :: Rep BufferMode -> BufferMode

Generic Newline

Associated Types

type Code Newline :: [[*]]

Methods

from :: Newline -> Rep Newline

to :: Rep Newline -> Newline

Generic NewlineMode

Associated Types

type Code NewlineMode :: [[*]]

Methods

from :: NewlineMode -> Rep NewlineMode

to :: Rep NewlineMode -> NewlineMode

Generic SeekMode

Associated Types

type Code SeekMode :: [[*]]

Methods

from :: SeekMode -> Rep SeekMode

to :: Rep SeekMode -> SeekMode

Generic MaskingState

Associated Types

type Code MaskingState :: [[*]]

Methods

from :: MaskingState -> Rep MaskingState

to :: Rep MaskingState -> MaskingState

Generic IOException

Associated Types

type Code IOException :: [[*]]

Methods

from :: IOException -> Rep IOException

to :: Rep IOException -> IOException

Generic ErrorCall

Associated Types

type Code ErrorCall :: [[*]]

Methods

from :: ErrorCall -> Rep ErrorCall

to :: Rep ErrorCall -> ErrorCall

Generic ArithException

Associated Types

type Code ArithException :: [[*]]

Methods

from :: ArithException -> Rep ArithException

to :: Rep ArithException -> ArithException

Generic All

Associated Types

type Code All :: [[*]]

Methods

from :: All -> Rep All

to :: Rep All -> All

Generic Any

Associated Types

type Code Any :: [[*]]

Methods

from :: Any -> Rep Any

to :: Rep Any -> Any

Generic CChar

Associated Types

type Code CChar :: [[*]]

Methods

from :: CChar -> Rep CChar

to :: Rep CChar -> CChar

Generic CSChar

Associated Types

type Code CSChar :: [[*]]

Methods

from :: CSChar -> Rep CSChar

to :: Rep CSChar -> CSChar

Generic CUChar

Associated Types

type Code CUChar :: [[*]]

Methods

from :: CUChar -> Rep CUChar

to :: Rep CUChar -> CUChar

Generic CShort

Associated Types

type Code CShort :: [[*]]

Methods

from :: CShort -> Rep CShort

to :: Rep CShort -> CShort

Generic CUShort

Associated Types

type Code CUShort :: [[*]]

Methods

from :: CUShort -> Rep CUShort

to :: Rep CUShort -> CUShort

Generic CInt

Associated Types

type Code CInt :: [[*]]

Methods

from :: CInt -> Rep CInt

to :: Rep CInt -> CInt

Generic CUInt

Associated Types

type Code CUInt :: [[*]]

Methods

from :: CUInt -> Rep CUInt

to :: Rep CUInt -> CUInt

Generic CLong

Associated Types

type Code CLong :: [[*]]

Methods

from :: CLong -> Rep CLong

to :: Rep CLong -> CLong

Generic CULong

Associated Types

type Code CULong :: [[*]]

Methods

from :: CULong -> Rep CULong

to :: Rep CULong -> CULong

Generic CLLong

Associated Types

type Code CLLong :: [[*]]

Methods

from :: CLLong -> Rep CLLong

to :: Rep CLLong -> CLLong

Generic CULLong

Associated Types

type Code CULLong :: [[*]]

Methods

from :: CULLong -> Rep CULLong

to :: Rep CULLong -> CULLong

Generic CFloat

Associated Types

type Code CFloat :: [[*]]

Methods

from :: CFloat -> Rep CFloat

to :: Rep CFloat -> CFloat

Generic CDouble

Associated Types

type Code CDouble :: [[*]]

Methods

from :: CDouble -> Rep CDouble

to :: Rep CDouble -> CDouble

Generic CPtrdiff

Associated Types

type Code CPtrdiff :: [[*]]

Methods

from :: CPtrdiff -> Rep CPtrdiff

to :: Rep CPtrdiff -> CPtrdiff

Generic CSize

Associated Types

type Code CSize :: [[*]]

Methods

from :: CSize -> Rep CSize

to :: Rep CSize -> CSize

Generic CWchar

Associated Types

type Code CWchar :: [[*]]

Methods

from :: CWchar -> Rep CWchar

to :: Rep CWchar -> CWchar

Generic CSigAtomic

Associated Types

type Code CSigAtomic :: [[*]]

Methods

from :: CSigAtomic -> Rep CSigAtomic

to :: Rep CSigAtomic -> CSigAtomic

Generic CClock

Associated Types

type Code CClock :: [[*]]

Methods

from :: CClock -> Rep CClock

to :: Rep CClock -> CClock

Generic CTime

Associated Types

type Code CTime :: [[*]]

Methods

from :: CTime -> Rep CTime

to :: Rep CTime -> CTime

Generic CUSeconds

Associated Types

type Code CUSeconds :: [[*]]

Methods

from :: CUSeconds -> Rep CUSeconds

to :: Rep CUSeconds -> CUSeconds

Generic CSUSeconds

Associated Types

type Code CSUSeconds :: [[*]]

Methods

from :: CSUSeconds -> Rep CSUSeconds

to :: Rep CSUSeconds -> CSUSeconds

Generic CIntPtr

Associated Types

type Code CIntPtr :: [[*]]

Methods

from :: CIntPtr -> Rep CIntPtr

to :: Rep CIntPtr -> CIntPtr

Generic CUIntPtr

Associated Types

type Code CUIntPtr :: [[*]]

Methods

from :: CUIntPtr -> Rep CUIntPtr

to :: Rep CUIntPtr -> CUIntPtr

Generic CIntMax

Associated Types

type Code CIntMax :: [[*]]

Methods

from :: CIntMax -> Rep CIntMax

to :: Rep CIntMax -> CIntMax

Generic CUIntMax

Associated Types

type Code CUIntMax :: [[*]]

Methods

from :: CUIntMax -> Rep CUIntMax

to :: Rep CUIntMax -> CUIntMax

Generic IOMode

Associated Types

type Code IOMode :: [[*]]

Methods

from :: IOMode -> Rep IOMode

to :: Rep IOMode -> IOMode

Generic Lexeme

Associated Types

type Code Lexeme :: [[*]]

Methods

from :: Lexeme -> Rep Lexeme

to :: Rep Lexeme -> Lexeme

Generic Number

Associated Types

type Code Number :: [[*]]

Methods

from :: Number -> Rep Number

to :: Rep Number -> Number

Generic GeneralCategory

Associated Types

type Code GeneralCategory :: [[*]]

Methods

from :: GeneralCategory -> Rep GeneralCategory

to :: Rep GeneralCategory -> GeneralCategory

HasDatatypeInfo [a]

Associated Types

type DatatypeInfoOf [a] :: DatatypeInfo

Methods

datatypeInfo :: proxy [a] -> DatatypeInfo (Code [a])

HasDatatypeInfo (Maybe a)

Associated Types

type DatatypeInfoOf (Maybe a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Maybe a) -> DatatypeInfo (Code (Maybe a))

HasDatatypeInfo (Complex a)

Associated Types

type DatatypeInfoOf (Complex a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Complex a) -> DatatypeInfo (Code (Complex a))

HasDatatypeInfo (Fixed a)

Associated Types

type DatatypeInfoOf (Fixed a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Fixed a) -> DatatypeInfo (Code (Fixed a))

HasDatatypeInfo (ArgOrder a)

Associated Types

type DatatypeInfoOf (ArgOrder a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (ArgOrder a) -> DatatypeInfo (Code (ArgOrder a))

HasDatatypeInfo (OptDescr a)

Associated Types

type DatatypeInfoOf (OptDescr a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (OptDescr a) -> DatatypeInfo (Code (OptDescr a))

HasDatatypeInfo (ArgDescr a)

Associated Types

type DatatypeInfoOf (ArgDescr a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (ArgDescr a) -> DatatypeInfo (Code (ArgDescr a))

HasDatatypeInfo (Dual a)

Associated Types

type DatatypeInfoOf (Dual a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Dual a) -> DatatypeInfo (Code (Dual a))

HasDatatypeInfo (Endo a)

Associated Types

type DatatypeInfoOf (Endo a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Endo a) -> DatatypeInfo (Code (Endo a))

HasDatatypeInfo (Sum a)

Associated Types

type DatatypeInfoOf (Sum a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Sum a) -> DatatypeInfo (Code (Sum a))

HasDatatypeInfo (Product a)

Associated Types

type DatatypeInfoOf (Product a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Product a) -> DatatypeInfo (Code (Product a))

HasDatatypeInfo (First a)

Associated Types

type DatatypeInfoOf (First a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (First a) -> DatatypeInfo (Code (First a))

HasDatatypeInfo (Last a)

Associated Types

type DatatypeInfoOf (Last a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Last a) -> DatatypeInfo (Code (Last a))

HasDatatypeInfo (Down a)

Associated Types

type DatatypeInfoOf (Down a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Down a) -> DatatypeInfo (Code (Down a))

HasDatatypeInfo (I a)

Associated Types

type DatatypeInfoOf (I a) :: DatatypeInfo

Methods

datatypeInfo :: proxy (I a) -> DatatypeInfo (Code (I a))

Generic [a]

Associated Types

type Code [a] :: [[*]]

Methods

from :: [a] -> Rep [a]

to :: Rep [a] -> [a]

Generic (Maybe a)

Associated Types

type Code (Maybe a) :: [[*]]

Methods

from :: Maybe a -> Rep (Maybe a)

to :: Rep (Maybe a) -> Maybe a

Generic (Complex a)

Associated Types

type Code (Complex a) :: [[*]]

Methods

from :: Complex a -> Rep (Complex a)

to :: Rep (Complex a) -> Complex a

Generic (Fixed a)

Associated Types

type Code (Fixed a) :: [[*]]

Methods

from :: Fixed a -> Rep (Fixed a)

to :: Rep (Fixed a) -> Fixed a

Generic (ArgOrder a)

Associated Types

type Code (ArgOrder a) :: [[*]]

Methods

from :: ArgOrder a -> Rep (ArgOrder a)

to :: Rep (ArgOrder a) -> ArgOrder a

Generic (OptDescr a)

Associated Types

type Code (OptDescr a) :: [[*]]

Methods

from :: OptDescr a -> Rep (OptDescr a)

to :: Rep (OptDescr a) -> OptDescr a

Generic (ArgDescr a)

Associated Types

type Code (ArgDescr a) :: [[*]]

Methods

from :: ArgDescr a -> Rep (ArgDescr a)

to :: Rep (ArgDescr a) -> ArgDescr a

Generic (Dual a)

Associated Types

type Code (Dual a) :: [[*]]

Methods

from :: Dual a -> Rep (Dual a)

to :: Rep (Dual a) -> Dual a

Generic (Endo a)

Associated Types

type Code (Endo a) :: [[*]]

Methods

from :: Endo a -> Rep (Endo a)

to :: Rep (Endo a) -> Endo a

Generic (Sum a)

Associated Types

type Code (Sum a) :: [[*]]

Methods

from :: Sum a -> Rep (Sum a)

to :: Rep (Sum a) -> Sum a

Generic (Product a)

Associated Types

type Code (Product a) :: [[*]]

Methods

from :: Product a -> Rep (Product a)

to :: Rep (Product a) -> Product a

Generic (First a)

Associated Types

type Code (First a) :: [[*]]

Methods

from :: First a -> Rep (First a)

to :: Rep (First a) -> First a

Generic (Last a)

Associated Types

type Code (Last a) :: [[*]]

Methods

from :: Last a -> Rep (Last a)

to :: Rep (Last a) -> Last a

Generic (Down a)

Associated Types

type Code (Down a) :: [[*]]

Methods

from :: Down a -> Rep (Down a)

to :: Rep (Down a) -> Down a

Generic (I a)

Associated Types

type Code (I a) :: [[*]]

Methods

from :: I a -> Rep (I a)

to :: Rep (I a) -> I a

HasDatatypeInfo (Either a b)

Associated Types

type DatatypeInfoOf (Either a b) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Either a b) -> DatatypeInfo (Code (Either a b))

HasDatatypeInfo (a, b)

Associated Types

type DatatypeInfoOf (a, b) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b) -> DatatypeInfo (Code (a, b))

HasDatatypeInfo (Proxy * t)

Associated Types

type DatatypeInfoOf (Proxy * t) :: DatatypeInfo

Methods

datatypeInfo :: proxy (Proxy * t) -> DatatypeInfo (Code (Proxy * t))

Generic (Either a b)

Associated Types

type Code (Either a b) :: [[*]]

Methods

from :: Either a b -> Rep (Either a b)

to :: Rep (Either a b) -> Either a b

Generic (a, b)

Associated Types

type Code (a, b) :: [[*]]

Methods

from :: (a, b) -> Rep (a, b)

to :: Rep (a, b) -> (a, b)

Generic (Proxy * t)

Associated Types

type Code (Proxy * t) :: [[*]]

Methods

from :: Proxy * t -> Rep (Proxy * t)

to :: Rep (Proxy * t) -> Proxy * t

HasDatatypeInfo (a, b, c)

Associated Types

type DatatypeInfoOf (a, b, c) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c) -> DatatypeInfo (Code (a, b, c))

HasDatatypeInfo (K * a b)

Associated Types

type DatatypeInfoOf (K * a b) :: DatatypeInfo

Methods

datatypeInfo :: proxy (K * a b) -> DatatypeInfo (Code (K * a b))

Generic (a, b, c)

Associated Types

type Code (a, b, c) :: [[*]]

Methods

from :: (a, b, c) -> Rep (a, b, c)

to :: Rep (a, b, c) -> (a, b, c)

Generic (K * a b)

Associated Types

type Code (K * a b) :: [[*]]

Methods

from :: K * a b -> Rep (K * a b)

to :: Rep (K * a b) -> K * a b

HasDatatypeInfo (a, b, c, d)

Associated Types

type DatatypeInfoOf (a, b, c, d) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d) -> DatatypeInfo (Code (a, b, c, d))

Generic (a, b, c, d)

Associated Types

type Code (a, b, c, d) :: [[*]]

Methods

from :: (a, b, c, d) -> Rep (a, b, c, d)

to :: Rep (a, b, c, d) -> (a, b, c, d)

HasDatatypeInfo (a, b, c, d, e)

Associated Types

type DatatypeInfoOf (a, b, c, d, e) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e) -> DatatypeInfo (Code (a, b, c, d, e))

HasDatatypeInfo ((:.:) * * f g p)

Associated Types

type DatatypeInfoOf ((* :.: *) f g p) :: DatatypeInfo

Methods

datatypeInfo :: proxy ((* :.: *) f g p) -> DatatypeInfo (Code ((* :.: *) f g p))

Generic (a, b, c, d, e)

Associated Types

type Code (a, b, c, d, e) :: [[*]]

Methods

from :: (a, b, c, d, e) -> Rep (a, b, c, d, e)

to :: Rep (a, b, c, d, e) -> (a, b, c, d, e)

Generic ((:.:) * * f g p)

Associated Types

type Code ((* :.: *) f g p) :: [[*]]

Methods

from :: (* :.: *) f g p -> Rep ((* :.: *) f g p)

to :: Rep ((* :.: *) f g p) -> (* :.: *) f g p

HasDatatypeInfo (a, b, c, d, e, f)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f) -> DatatypeInfo (Code (a, b, c, d, e, f))

Generic (a, b, c, d, e, f)

Associated Types

type Code (a, b, c, d, e, f) :: [[*]]

Methods

from :: (a, b, c, d, e, f) -> Rep (a, b, c, d, e, f)

to :: Rep (a, b, c, d, e, f) -> (a, b, c, d, e, f)

HasDatatypeInfo (a, b, c, d, e, f, g)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g) -> DatatypeInfo (Code (a, b, c, d, e, f, g))

Generic (a, b, c, d, e, f, g)

Associated Types

type Code (a, b, c, d, e, f, g) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g) -> Rep (a, b, c, d, e, f, g)

to :: Rep (a, b, c, d, e, f, g) -> (a, b, c, d, e, f, g)

HasDatatypeInfo (a, b, c, d, e, f, g, h)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h))

Generic (a, b, c, d, e, f, g, h)

Associated Types

type Code (a, b, c, d, e, f, g, h) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h) -> Rep (a, b, c, d, e, f, g, h)

to :: Rep (a, b, c, d, e, f, g, h) -> (a, b, c, d, e, f, g, h)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i))

Generic (a, b, c, d, e, f, g, h, i)

Associated Types

type Code (a, b, c, d, e, f, g, h, i) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i) -> Rep (a, b, c, d, e, f, g, h, i)

to :: Rep (a, b, c, d, e, f, g, h, i) -> (a, b, c, d, e, f, g, h, i)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j))

Generic (a, b, c, d, e, f, g, h, i, j)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j) -> Rep (a, b, c, d, e, f, g, h, i, j)

to :: Rep (a, b, c, d, e, f, g, h, i, j) -> (a, b, c, d, e, f, g, h, i, j)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k))

Generic (a, b, c, d, e, f, g, h, i, j, k)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k) -> Rep (a, b, c, d, e, f, g, h, i, j, k)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k) -> (a, b, c, d, e, f, g, h, i, j, k)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l))

Generic (a, b, c, d, e, f, g, h, i, j, k, l)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l) -> (a, b, c, d, e, f, g, h, i, j, k, l)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m) -> (a, b, c, d, e, f, g, h, i, j, k, l, m)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28)

HasDatatypeInfo (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29)

Associated Types

type DatatypeInfoOf (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29) :: DatatypeInfo

Methods

datatypeInfo :: proxy (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29) -> DatatypeInfo (Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29))

Generic (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29)

Associated Types

type Code (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29) :: [[*]]

Methods

from :: (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29) -> Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29)

to :: Rep (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29) -> (a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z, t26, t27, t28, t29)