Generics.SOP.Metadata

Description

Metadata about what a datatype looks like

In generics-sop, the metadata is completely independent of the main universe. Many generic functions will use this metadata, but other don't, and yet others might need completely different metadata.

This module defines a datatype to represent standard metadata, i.e., names of the datatype, its constructors, and possibly its record selectors. Metadata descriptions are in general GADTs indexed by the code of the datatype they're associated with, so matching on the metadata will reveal information about the shape of the datatype.

Documentation

type Fixity = Int

The fixity of an infix constructor.

data ConstructorInfo :: [Type] -> Type where

Metadata for a single constructor.

This is indexed by the product structure of the constructor components.

Constructors

Constructor :: SListI xs => ConstructorName -> ConstructorInfo xs

Infix :: ConstructorName -> Associativity -> Fixity -> ConstructorInfo '[x, y]

Record :: SListI xs => ConstructorName -> NP FieldInfo xs -> ConstructorInfo xs

data DatatypeInfo :: [[Type]] -> Type where

Metadata for a datatype.

A value of type DatatypeInfo c contains the information about a datatype that is not contained in Code c. This information consists primarily of the names of the datatype, its constructors, and possibly its record selectors.

The constructor indicates whether the datatype has been declared using newtype or not.

Constructors

ADT :: ModuleName -> DatatypeName -> NP ConstructorInfo xss -> POP StrictnessInfo xss -> DatatypeInfo xss

Newtype :: ModuleName -> DatatypeName -> ConstructorInfo '[x] -> DatatypeInfo '['[x]]

data FieldInfo :: Type -> Type where

For records, this functor maps the component to its selector name.

Constructors

FieldInfo :: FieldName -> FieldInfo a

type DatatypeName = String

The name of a datatype.

type ModuleName = String

The name of a module.

type ConstructorName = String

The name of a data constructor.

type FieldName = String

The name of a field / record selector.

data StrictnessInfo :: Type -> Type where

Metadata for strictness information of a field.

Indexed by the type of the field.

Since: 0.4.0.0

Constructors

StrictnessInfo :: SourceUnpackedness -> SourceStrictness -> DecidedStrictness -> StrictnessInfo a

datatypeName :: DatatypeInfo xss -> DatatypeName

The name of a datatype (or newtype).

Since: 0.2.3.0

moduleName :: DatatypeInfo xss -> ModuleName

The module name where a datatype is defined.

Since: 0.2.3.0

constructorName :: ConstructorInfo xs -> ConstructorName

The name of a constructor.

Since: 0.2.3.0

constructorInfo :: DatatypeInfo xss -> NP ConstructorInfo xss

The constructor info for a datatype (or newtype).

Since: 0.2.3.0

fieldName :: FieldInfo a -> FieldName

The name of a field.

Since: 0.2.3.0

re-exports

data Associativity

Datatype to represent the associativity of a constructor

Constructors

LeftAssociative

RightAssociative

NotAssociative

data DecidedStrictness

The strictness that GHC infers for a field during compilation. Whereas there are nine different combinations of SourceUnpackedness and SourceStrictness, the strictness that GHC decides will ultimately be one of lazy, strict, or unpacked. What GHC decides is affected both by what the user writes in the source code and by GHC flags. As an example, consider this data type:

data E = ExampleConstructor {-# UNPACK #-} !Int !Int Int

If compiled without optimization or other language extensions, then the fields of ExampleConstructor will have DecidedStrict, DecidedStrict, and DecidedLazy, respectively.

If compiled with -XStrictData enabled, then the fields will have DecidedStrict, DecidedStrict, and DecidedStrict, respectively.

If compiled with -O2 enabled, then the fields will have DecidedUnpack, DecidedStrict, and DecidedLazy, respectively.

Since: base-4.9.0.0

Constructors

DecidedLazy

DecidedStrict

DecidedUnpack

data SourceStrictness

The strictness of a field as the user wrote it in the source code. For example, in the following data type:

data E = ExampleConstructor Int ~Int !Int

The fields of ExampleConstructor have NoSourceStrictness, SourceLazy, and SourceStrict, respectively.

Since: base-4.9.0.0

Constructors

NoSourceStrictness

SourceLazy

SourceStrict

data SourceUnpackedness

The unpackedness of a field as the user wrote it in the source code. For example, in the following data type:

data E = ExampleConstructor     Int
           {-# NOUNPACK #-} Int
           {-#   UNPACK #-} Int

The fields of ExampleConstructor have NoSourceUnpackedness, SourceNoUnpack, and SourceUnpack, respectively.

Since: base-4.9.0.0

Constructors

NoSourceUnpackedness

SourceNoUnpack

SourceUnpack