Generics.SOP.Metadata

Description

Metadata about what a datatype looks like

In generics-sop, the metadata is completely independent of the main universe. Many generic functions will use this metadata, but other don't, and yet others might need completely different metadata.

This module defines a datatype to represent standard metadata, i.e., names of the datatype, its constructors, and possibly its record selectors. Metadata descriptions are in general GADTs indexed by the code of the datatype they're associated with, so matching on the metadata will reveal information about the shape of the datatype.

Documentation

data DatatypeInfo :: [[*]] -> * where

Metadata for a datatype.

A value of type DatatypeInfo c contains the information about a datatype that is not contained in Code c. This information consists primarily of the names of the datatype, its constructors, and possibly its record selectors.

The constructor indicates whether the datatype has been declared using newtype or not.

Constructors

ADT :: ModuleName -> DatatypeName -> NP ConstructorInfo xss -> DatatypeInfo xss

Newtype :: ModuleName -> DatatypeName -> ConstructorInfo '[x] -> DatatypeInfo '['[x]]

moduleName :: DatatypeInfo xss -> ModuleName

The module name where a datatype is defined.

Since: 0.2.3.0

datatypeName :: DatatypeInfo xss -> DatatypeName

The name of a datatype (or newtype).

Since: 0.2.3.0

constructorInfo :: DatatypeInfo xss -> NP ConstructorInfo xss

The constructor info for a datatype (or newtype).

Since: 0.2.3.0

data ConstructorInfo :: [*] -> * where

Metadata for a single constructors.

This is indexed by the product structure of the constructor components.

Constructors

Constructor :: SListI xs => ConstructorName -> ConstructorInfo xs

Infix :: ConstructorName -> Associativity -> Fixity -> ConstructorInfo '[x, y]

Record :: SListI xs => ConstructorName -> NP FieldInfo xs -> ConstructorInfo xs

constructorName :: ConstructorInfo xs -> ConstructorName

The name of a constructor.

Since: 0.2.3.0

data FieldInfo :: * -> * where

For records, this functor maps the component to its selector name.

Constructors

FieldInfo :: FieldName -> FieldInfo a

fieldName :: FieldInfo a -> FieldName

The name of a field.

Since: 0.2.3.0

type DatatypeName = String

The name of a datatype.

type ModuleName = String

The name of a module.

type ConstructorName = String

The name of a data constructor.

type FieldName = String

The name of a field / record selector.

type Fixity = Int

The fixity of an infix constructor.

re-exports

data Associativity :: *

Datatype to represent the associativity of a constructor

Constructors

LeftAssociative

RightAssociative

NotAssociative