Generics.SOP.Type.Metadata

Description

Type-level metadata

This module provides datatypes (to be used promoted) that can represent the metadata of Haskell datatypes on the type level.

We do not reuse the term-level metadata types, because these are GADTs that incorporate additional invariants. We could (at least in GHC 8) impose the same invariants on the type level as well, but some tests have revealed that the resulting type are rather inconvenient to work with.

So we use simple datatypes to represent the type-level metadata, even if this means that some invariants are not explicitly captured.

We establish a relation between the term- and type-level versions of the metadata by automatically computing the term-level version from the type-level version.

As we now have two versions of metadata (term-level and type-level) with very similar, yet slightly different datatype definitions, the names between the modules clash, and this module is recommended to be imported qualified when needed.

The interface exported by this module is still somewhat experimental.

Since: 0.3.0.0

Documentation

data DatatypeInfo

Metadata for a datatype (to be used promoted).

A type of kind DatatypeInfo contains meta-information about a datatype that is not contained in its code. This information consists primarily of the names of the datatype, its constructors, and possibly its record selectors.

The constructor indicates whether the datatype has been declared using newtype or not.

Since: 0.3.0.0

Constructors

ADT ModuleName DatatypeName [ConstructorInfo]
Standard algebraic datatype

Newtype ModuleName DatatypeName ConstructorInfo
Newtype

data ConstructorInfo

Metadata for a single constructors (to be used promoted).

Since: 0.3.0.0

Constructors

Constructor ConstructorName
Normal constructor

Infix ConstructorName Associativity Fixity
Infix constructor

Record ConstructorName [FieldInfo]
Record constructor

data FieldInfo

Metadata for a single record field (to be used promoted).

Since: 0.3.0.0

Constructors

FieldInfo FieldName

type DatatypeName = Symbol

The name of a datatype.

type ModuleName = Symbol

The name of a module.

type ConstructorName = Symbol

The name of a data constructor.

type FieldName = Symbol

The name of a field / record selector.

type Fixity = Nat

The fixity of an infix constructor.

class DemoteDatatypeInfo (x :: DatatypeInfo) (xss :: [[*]]) where

Class for computing term-level datatype information from type-level datatype information.

Since: 0.3.0.0

Minimal complete definition

demoteDatatypeInfo

Methods

demoteDatatypeInfo :: proxy x -> DatatypeInfo xss

Given a proxy of some type-level datatype information, return the corresponding term-level information.

Since: 0.3.0.0

class DemoteConstructorInfos (cs :: [ConstructorInfo]) (xss :: [[*]]) where

Class for computing term-level constructor information from type-level constructor information.

Since: 0.3.0.0

Minimal complete definition

demoteConstructorInfos

Methods

demoteConstructorInfos :: proxy cs -> NP ConstructorInfo xss

Given a proxy of some type-level constructor information, return the corresponding term-level information as a product.

Since: 0.3.0.0

class DemoteConstructorInfo (x :: ConstructorInfo) (xs :: [*]) where

Class for computing term-level constructor information from type-level constructor information.

Since: 0.3.0.0

Minimal complete definition

demoteConstructorInfo

Methods

demoteConstructorInfo :: proxy x -> ConstructorInfo xs

Given a proxy of some type-level constructor information, return the corresponding term-level information.

Since: 0.3.0.0

class SListI xs => DemoteFieldInfos (fs :: [FieldInfo]) (xs :: [*]) where

Class for computing term-level field information from type-level field information.

Since: 0.3.0.0

Minimal complete definition

demoteFieldInfos

Methods

demoteFieldInfos :: proxy fs -> NP FieldInfo xs

Given a proxy of some type-level field information, return the corresponding term-level information as a product.

Since: 0.3.0.0

class DemoteFieldInfo (x :: FieldInfo) (a :: *) where

Class for computing term-level field information from type-level field information.

Since: 0.3.0.0

Minimal complete definition

demoteFieldInfo

Methods

demoteFieldInfo :: proxy x -> FieldInfo a

Given a proxy of some type-level field information, return the corresponding term-level information.

Since: 0.3.0.0

class DemoteAssociativity (a :: Associativity) where

Class for computing term-level associativity information from type-level associativity information.

Since: 0.3.0.0

Minimal complete definition

demoteAssociativity

Methods

demoteAssociativity :: proxy a -> Associativity

Given a proxy of some type-level associativity information, return the corresponding term-level information.

Since: 0.3.0.0

re-exports

data Associativity :: *

Datatype to represent the associativity of a constructor

Constructors

LeftAssociative

RightAssociative

NotAssociative